+++ app/src/main/java/com/rdd/finy/helpers/AddCalculator.kt
@@ -3,9 +3,9 @@ package com.rdd.finy.helpers
 import com.rdd.finy.app.models.Wallet
 
 class AddCalculator(
-    override var userBalance: Int,
-    override var userConfigWallets: HashMap<Wallet, Int>,
-    override var otherWallets: List<Wallet>
+        override var userBalance: Long,
+        override var userConfigWallets: HashMap<Wallet, Long>,
+        override var otherWallets: List<Wallet>
 ) : CalculatorBeverage() {
 
     override fun getCalculatedResult(): List<Wallet> {
@@ -14,14 +14,14 @@ class AddCalculator(
             calculateUserConfigWallets()
         if (otherWallets.isNotEmpty())
             calculateOtherWallets()
-        if (userBalance != 0)
+        if (userBalance != 0.toLong())
             setupExtraWallet()
 
         return otherWallets + userConfigWallets.keys
     }
 
     override fun calculateUserConfigWallets() {
-        var diff: Int
+        var diff: Long
 
         for (item in userConfigWallets) {
 
@@ -45,8 +45,8 @@ class AddCalculator(
 
     override fun calculateOtherWallets() {
         var partsCount = otherWallets.size
-        var part: Int
-        var diff: Int
+        var part: Long
+        var diff: Long
 
         for (wallet in otherWallets) {
