+++ app/src/main/java/com/rdd/finy/helpers/CalculatorBeverage.kt
@@ -6,8 +6,8 @@ abstract class CalculatorBeverage {
 
     private val TAG = CalculatorBeverage::class.java.simpleName
 
-    abstract var userBalance: Int
-    abstract var userConfigWallets: HashMap<Wallet, Int>
+    abstract var userBalance: Long
+    abstract var userConfigWallets: HashMap<Wallet, Long>
     abstract var otherWallets: List<Wallet>
 
     abstract fun getCalculatedResult(): List<Wallet>
